+++ lib/services/user.dart
@@ -40,6 +40,7 @@ import 'package:Okuna/models/reactions_emoji_count_list.dart';
 import 'package:Okuna/models/posts_list.dart';
 import 'package:Okuna/models/top_post.dart';
 import 'package:Okuna/models/top_posts_list.dart';
+import 'package:Okuna/models/trending_posts_list.dart';
 import 'package:Okuna/models/user.dart';
 import 'package:Okuna/models/user_invite.dart';
 import 'package:Okuna/models/user_invites_list.dart';
@@ -483,13 +484,17 @@ class UserService {
     return TopPostsList.fromJson(json.decode(response.body));
   }
 
-  Future<PostsList> getTrendingPosts() async {
+  Future<TrendingPostsList> getTrendingPosts({int maxId, int minId, int count}) async {
     HttpieResponse response =
-        await _postsApiService.getTrendingPosts(authenticatedRequest: true);
+        await _postsApiService.getTrendingPosts(
+            maxId: maxId,
+            minId: minId,
+            count: count,
+            authenticatedRequest: true);
 
     _checkResponseIsOk(response);
 
-    return PostsList.fromJson(json.decode(response.body));
+    return TrendingPostsList.fromJson(json.decode(response.body));
   }
 
   Future<PostsList> getTimelinePosts(